Match all pages on GitHub
When starting navigation from the notifications page, upon going to the issue page, nothing happens because the page URL isn't actually changed because of how GitHub loads client-side. 

This change matches on all URLs so the script is loaded on all pages ready for when the user navigates to the issue page.